Move marshaler to stdoutlog exporter
